remove webhook api error switch and implement error mapping with error switch
