+++ crates/router/src/core/errors.rs
@@ -1,5 +1,6 @@
 pub mod api_error_response;
 pub mod error_handlers;
+pub mod transformers;
 pub mod utils;
 
 use std::fmt::Display;
